putting appflow lambda code inline
